# app/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trash Fairy — API",
    "version": "1.1.0",
    "summary": "Otwarte API /api/v1 (tylko odczyt, CORS) oraz API aplikacji: kosze, zgłoszenia, trasa kierowcy, dashboard miasta, urządzenia IoT.",
    "description": "Dane operacyjne są syntetyczne (meta.synthetic = true). Pozycje koszy: © OpenStreetMap contributors (ODbL 1.0). Bez kluczy, CORS *, odpowiedzi cache'owane 60 s.",
    "license": {
      "name": "ODbL 1.0 (pozycje z OSM)",
      "identifier": "ODbL-1.0"
    }
  },
  "servers": [
    {
      "url": "https://trashfairy.twapp.pl"
    }
  ],
  "paths": {
...
    "/api/urzadzenia": {
      "get": {
        "tags": [
          "Urządzenia (IoT)"
        ],
        "summary": "Urządzenia na koszach: bateria, status, odczyty (filtry typ, dzielnica, status)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "typ",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dzielnica",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/urzadzenia/{id}": {
      "get": {
        "tags": [
          "Urządzenia (IoT)"
        ],
        "summary": "Jedno urządzenie z serią odczytów z 7 dni",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Brak urządzenia",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/odczyty": {
      "post": {
        "tags": [
          "Urządzenia (IoT)"
        ],
        "summary": "Odczyt z urządzenia (czujnik: zapełnienie; panel: sygnał życia). Nagłówek X-Token-Urzadzenia = token wgrany przy montażu",
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Wartość spoza zakresu 0–100",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Nieznane urządzenie albo zły token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Częściej niż raz na 10 s",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Token-Urzadzenia",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "numer_seryjny": "TF-US-000133",
                "zapelnienie": 64,
                "autotest_ok": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "const": false
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}